### 2019-02-18 eccc17e
Including paper
