### 2019-02-18 #1259 @mikldk
I totally agree. I forgot to change `.gitignore` (mikldk/mitolina@eccc17e3ada22fda0e920732523beb63d39c1ec9)
