+++ paper.bib
@@ -1,0 +1,135 @@
+@article{AndersenPLOSGEN2017,
+	Author = {Mikkel M Andersen and David J Balding},
+	Journal = {PLOS Genetics},
+	Pages = {e1007028},
+	Title = {{How convincing is a matching Y-chromosome profile?}},
+	Volume = {13},
+	Number = {11},
+	Year = {2017},
+	doi = {10.1371/journal.pgen.1007028}
+}
+
+
+@article{AndersenPLOSGENmtDNA2018,
+	Author = {Mikkel M Andersen and David J Balding},
+	Journal = {PLOS Genetics},
+	Pages = {e1007774},
+	Title = {{How many individuals share a mitochondrial genome?}},
+	Volume = {14},
+	Number = {11},
+	Year = {2018},
+	doi = {10.1371/journal.pgen.1007774}
+}
+
+@article{AndersenCoalescent2013,
+	Author = {Mikkel M Andersen and Amke Caliebe and Arne Jochens and Sascha Willuweit and Michael Krawczak},
+	Issue = {2},
+	Journal = {Forensic Science International: Genetics},
+	Pages = {264-271},
+	Title = {{Estimating trace-suspect match probabilities for singleton Y-STR haplotypes using coalescent theory}},
+	Volume = {7},
+	Year = {2013},
+	doi = {10.1016/j.fsigen.2012.11.004}
+}
+
+@article{AndersenDisclap2013,
+	Author = {Mikkel M Andersen and Poul S Eriksen and Niels Morling},
+	Journal = {Journal of Theoretical Biology},
+	Pages = {39--51},
+	Title = {{The discrete Laplace exponential family and estimation of Y-STR haplotype frequencies}},
+	Volume = {329},
+	Year = {2013},
+	doi = {10.1016/j.jtbi.2013.03.009}
+}
+
+@manual{R,
+	Address = {Vienna, Austria},
+	Author = {{R Core Team}},
+	Organization = {R Foundation for Statistical Computing},
+	Title = {R: A Language and Environment for Statistical Computing},
+	Url = {https://www.R-project.org/},
+	Year = {2018},
+	Bdsk-Url-1 = {https://www.R-project.org/}}
+
+@Article{Rcpp,
+  title = {{Extending \textit{R} with \textit{C++}: A Brief Introduction to \textit{Rcpp}}},
+  author = {D Eddelbuettel and J J Balamuta},
+  journal = {PeerJ Preprints},
+  year = {2017},
+  month = {aug},
+  volume = {5},
+  pages = {e3188v1},
+  issn = {2167-9843},
+  url = {https://doi.org/10.7287/peerj.preprints.3188v1},
+  doi = {10.7287/peerj.preprints.3188v1},
+}
+
+@book{DJBwoe2,
+	Author = {Steele, CD and Balding, DJ},
+	Date-Added = {2017-01-19 09:44:04 +0000},
+	Date-Modified = {2017-01-25 23:52:56 +0000},
+	Edition = {2nd ed.},
+	Publisher = {Wiley},
+	Title = {Weight of evidence for forensic {DNA} profiles},
+	Year = 2015,	
+	doi = {10.1002/9780470867693}
+}
+
+@book{Butler2009,
+	Author = {John M. Butler},
+	Isbn = {978-0123749994},
+	Publisher = {Academic Press},
+	Title = {Fundamentals of Forensic DNA Typing},
+	Year = {2009}}
+
+
+@article{Soares2009,
+  title={{Correcting for Purifying Selection: An Improved Human Mitochondrial Molecular Clock}},
[…]
