### 2019-02-18 #1259 @mikldk
Thanks, @pjotrp. I give it a thought. Although not in the centre of my field, @afrubin and @ajank have previously reviewed related software (#684); maybe they are ready again?
