### 2019-02-19 #1266 @whedon
[REVIEW]: mitolina: MITOchondrial LINeage Analysis
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/mitolina" target ="_blank">https://github.com/mikldk/mitolina</a>
**Version:** v0.0.1
