### 2019-02-19 #1266 @lpantano
Hi @mikldk,

We are starting the review process in this issue tracker.

To expedite the review process do you mind going through the above list of checkboxes and make sure they can be ticked (you can't tick them). Also, check the PDF output carefully.

Please, Ping us here when you are done.

Cheers

Lorena
