### 2019-02-19 #1259 @danielskatz
FWIW, @lpantano, you shouldn't need to use the "magic-word" - just ```@whedon start review``` should work
