### 2019-02-19 #1266 @mikldk
Hi @lpantano. Ping: I went though the list and I seem to have addressed all.
