### 2019-02-20 mikldk/mitolina#2 @jdeligt
Please make an official release
The current repository does not have an official release.
Please make one so that people have a stable version to download/use.
