### 2019-02-20 #1266 @jdeligt
Hi @mikldk I just raised a minor issue on not having a release in the repo. Could you please look into this while I continue with the rest of the review.
https://github.com/mikldk/mitolina/issues/2
