### 2019-02-20 mikldk/mitolina#3 @jdeligt
Vignette not loading
In a clean R installation running the vignette commands after remotes::install_github("mikldk/mitolina") results in an error. "no vignettes found". This might have to do with the install paths, please try to reproduce and adjust documentation/install accordingly.

When trying to knit the vignette it misses the ggraph (and underlying dependencies), it might be worth going over your complete dependency tree, i.e. not just the package but also the documentation and list these in your README and included them in the installer.
