### 2019-02-20 #1266 @jdeligt
Now hit a bigger issue: https://github.com/mikldk/mitolina/issues/3.
I have to manually install dependencies upon which relies on a 'non standard library': libudunits2.so
