### 2019-02-20 #1266 @jdeligt
@mikldk please let me know when you've looked at this so I can check the last steps. Apart from the software side of things I would like to hear your take on the Biparental inheritance published in November: https://doi.org/10.1073/pnas.1810946115 and how your software deals with this. I think its an important thing to note for potential users especially if you're targeting forensics where edge cases matter a lot.
