### 2019-02-20 595ef6a
Resolving #3. Also adapt vignette to compile on systems without ggraph.
