+++ vignettes/introduction.Rmd
@@ -117,25 +117,29 @@ We can then try to plot the same pedigree again, this time with information abou
 plot(ped, num_vars = TRUE)
 ```
 
-Plotting with the `igraph` plot quickly becomes difficult to customise. Therefore extensions to the `tidygraph`/`ggraph` packages is provided:
+Plotting with the `igraph` plot quickly becomes difficult to customise. Therefore extensions to the `tidygraph`/`ggraph` packages is provided (requires installation of the `ggraph` package):
 
 ```{r, fig.height=8, fig.width=12}
-library(tidygraph)
-library(ggraph)
-
-g <- as_tbl_graph(peds) %>% 
-  activate(nodes) %>% 
-  filter(ped_id == ped_id_endgen) %>% 
-  mutate(num_vars = sapply(haplotype, sum))
-
-p <- ggraph(g, layout = 'tree') +
-  geom_edge_link() +
-  geom_node_point(aes(color = sex, fill = num_vars, shape = sex), size = 12, stroke = 2) +
-  geom_node_text(aes(label = paste0(name, "\n", num_vars)), size = 4, color = "white") +
-  theme_graph(base_family = "") +
-  scale_shape_manual(NULL, values = c("Female" = 21, "Male" = 22)) +
-  scale_color_manual(NULL, values = c("Female" = "black", "Male" = "red"))
-p
+if (require("ggraph", quietly = TRUE)) {
+  library(tidygraph)
+  library(ggraph)
+  
+  g <- as_tbl_graph(peds) %>% 
+    activate(nodes) %>% 
+    filter(ped_id == ped_id_endgen) %>% 
+    mutate(num_vars = sapply(haplotype, sum))
+  
+  p <- ggraph(g, layout = 'tree') +
+    geom_edge_link() +
+    geom_node_point(aes(color = sex, fill = num_vars, shape = sex), size = 12, stroke = 2) +
+    geom_node_text(aes(label = paste0(name, "\n", num_vars)), size = 4, color = "white") +
+    theme_graph(base_family = "") +
+    scale_shape_manual(NULL, values = c("Female" = 21, "Male" = 22)) +
+    scale_color_manual(NULL, values = c("Female" = "black", "Male" = "red"))
+  p
+} else {
+  cat("The `ggraph` package is not installed, so the figure is not shown.")
+}
 ```
 
 Here, the top number is the id and the number below is the number of variants in the individual's mitogenome.
@@ -282,19 +286,23 @@ table(sapply(haps, sum))
 And the same pedigree with new mitogenomes, now where number of variants is a factor to get a discrete fill colour scale:
 
 ```{r, fig.height=8, fig.width=12}
-g <- as_tbl_graph(peds) %>% 
-  activate(nodes) %>% 
-  filter(ped_id == ped_id_endgen) %>% 
-  mutate(num_vars = sapply(haplotype, sum))
-
-p <- ggraph(g, layout = 'tree') +
-  geom_edge_link() +
-  geom_node_point(aes(color = sex, fill = factor(num_vars), shape = sex), size = 12, stroke = 2, show.legend = FALSE) +
-  geom_node_text(aes(label = paste0(name, "\n", num_vars)), size = 4, color = "white") +
-  theme_graph(base_family = "") +
-  scale_shape_manual(NULL, values = c("Female" = 21, "Male" = 22)) +
-  scale_color_manual(NULL, values = c("Female" = "black", "Male" = "red"))
-p
+if (require("ggraph", quietly = TRUE)) {
+  g <- as_tbl_graph(peds) %>% 
+    activate(nodes) %>% 
+    filter(ped_id == ped_id_endgen) %>% 
[…]
