### 2019-02-20 #1266 @mikldk
@jdeligt:
Thanks for all your inputs!

1) Regarding a release (mikldk/mitolina#2): I thought I should wait until after potential acceptance. But I now made one [here](https://github.com/mikldk/mitolina/releases).
2) Vignette not loading (mikldk/mitolina#3): I have now updated `README.md` to compile these by default (`remotes::install_github("mikldk/mitolina", build_opts = c("--no-resave-data", "--no-manual"))`). Thanks.
3) Dependencies: I have now adapted the vignette so that it checks whether the suggested `ggraph` is installed. This ought to remove the "'non standard library': libudunits2.so" warning. Thanks.
4) Regarding [10.1073/pnas.1810946115](https://doi.org/10.1073/pnas.1810946115): There is still very much debate about this. See e.g. [10.1073/pnas.1820533116](https://doi.org/10.1073/pnas.1820533116). Still, the general belief is that mtDNA is maternally inherited.

Please let me know if I can assist further.
