### 2019-02-20 #1266 @jdeligt
@mikldk thanks fro the quick response and resolving the issues. As far as the software/paper is concerned I'm happy to accept this and move it forward. I would personally add a (small) disclaimer about the fact that this software operates under the maternal only model, just so that users are aware of it.
@lpantano please move this paper forward.
