### 2019-02-20 ae28a61
Add note about maternal inheritance only model
