### 2019-02-20 #1266 @mikldk
Thanks, @jdeligt, @pjotrp and @lpantano. I have now added the note "This software operates under the maternal inheritance only model, i.e. that mtDNA is only passed on by mothers to children." in both `README.md` and `paper.md`. I have also deleted the old release and made a new: <https://github.com/mikldk/mitolina/releases>. I have also added Zenodo, and this new release has DOI: <https://doi.org/10.5281/zenodo.2574140>.
