### 2019-02-20 #1266 @mikldk
@lpantano , yes indeed. I just wanted to check the proof before I wrote you :-). The note was added in mikldk/mitolina@ae28a61f9c137fe1574ba0d59f320b6ff6414069.
