### 2019-02-20 #1266 @lpantano
Hi @pjotrp,

all is good, should you do the deposit, or I do?

Thanks!
