### 2019-02-20 #1266 @labarba
(We don't remove the "review" label.)
