### 2019-02-20 #1266 @labarba
:wave @mikldk — Could you edit the metadata of your Zenodo archive so it matches the title of the JOSS paper? Thanks.
