### 2019-02-20 #1266 @labarba
You may also want to add case protection with `{}` to the bibtex entry for Butler (2009), so you get _DNA_ instead of _dna_.
