### 2019-02-20 #1266 @lpantano
@labarba, is the metadata to change the one that is saying now: `mikldk/mitolina: v0.0.1`? (just to learn for the next time) and (sorry about the label).
