### 2019-02-20 #1266 @labarba
Yes: the title of the Zenodo archive matching the title of the JOSS paper.
