### 2019-02-20 #1266 @labarba
(Authors often need to be manually edited, too, as Zenodo automatically lists everyone who has a commit.)
