### 2019-02-21 #1266 @mikldk
@lpantano @labarba I'll look at Zenodo metadata. And the reference. And recompile paper. I guess I should then delete Github release and make a new afterwards?
