+++ paper.bib
@@ -79,7 +79,7 @@
 	Author = {John M. Butler},
 	Isbn = {978-0123749994},
 	Publisher = {Academic Press},
-	Title = {Fundamentals of Forensic DNA Typing},
+	Title = {{Fundamentals of Forensic DNA Typing}},
 	Year = {2009}}
 
 
