### 2019-02-21 #1266 @mikldk
@lpantano @labarba : I have updated `paper.bib`, deleted the old release and created a new. The new DOI at Zenodo is [10.5281/zenodo.2574400](https://doi.org/10.5281/zenodo.2574400). I have changed the Zenodo title at <https://zenodo.org/record/2574400>. I hope I have done everything right?
